+++ src/Event/UserDataEvent.php
@@ -13,12 +13,8 @@ class UserDataEvent extends Event
         $this->data = $data;
     }
 
-    public function getData()
+    public function getData(): array
     {
         return $this->data;
-    }
-    public function setData(array $data)
-    {
-        $this->data = $data;
     }
 }