+++ src/Exception/TokenException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace ItkDev\Adgangsstyring\Exception;
+
+class TokenException extends \Exception
+{
+}