+++ CHANGELOG.md
@@ -10,11 +10,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Added
 
 - Controller
-- Start, Commit and UserData events
+- StartEvent, CommitEvent and UserDataEvent
 - README
 - LICENSE
 - PHP_CodeSniffer
 - markdownlint
 - This CHANGELOG file to hopefully serve as an evolving example of
   a standardized open source project CHANGELOG.
-  
+- PHPUnit
+- ControllerTest